+++ packages/headers/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This is the changelog for [`headers`](https://github.com/mjackson/remix-the-web/tree/main/packages/headers). It follows [semantic versioning](https://semver.org/).
 
-## HEAD
+## v0.10.0 (2025-01-27)
 
 This release contains several improvements to `Cookie` that bring it more in line with other headers like `Accept`, `AcceptEncoding`, and `AcceptLanguage`.